# docs/api/api.yaml
openapi: "3.0.0"
info:
  title: "Wutsi Cart API"
  description: API for managing Online Cart
  version: "1.0.0"

servers:
  - url: https://wutsi-cart-server-test.herokuapp.com
    description: Sandbox
  - url: https://wutsi-cart-server-prod.herokuapp.com
    description: Production

paths:
  /v1/carts/search:
    post:
      operationId: "search-carts"
      description: Search Cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchCartRequest'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchCartResponse'
        403:
          description: The caller is not authorized to call this endpoint
      tags:
        - Product
      security:
        - api_key: [ 'cart-read' ]

  /v1/carts/{merchant-id}:
    get:
      operationId: "get-cart"
      description: Return the Cart of a user associated with a given marchant
      parameters:
        - in: path
          name: merchant-id
          description: ID of the merchant
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCartResponse'
      tags:
        - Cart
      security:
        - api_key: [ 'cart-read' ]

  /v1/carts/{merchant-id}/products:
    post:
      operationId: "add-product"
      description: Add a Product into a cart
      parameters:
        - in: path
          name: merchant-id
          description: ID of the merchant
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProductRequest'
      responses:
        200:
          description: Success
      tags:
        - Cart
      security:
        - api_key: [ 'cart-manage' ]

    delete:
      operationId: "empty-cart"
      description: Empty a cart
      parameters:
        - in: path
          name: merchant-id
          description: ID of the merchant
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
      tags:
        - Cart
      security:
        - api_key: [ 'cart-read' ]

  /v1/carts/{merchant-id}/products/{product-id}:
    post:
      operationId: "update-product"
      description: Update a Product into a cart
      parameters:
        - in: path
          name: merchant-id
          description: ID of the merchant
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: product-id
          description: ID of the product to add
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        200:
          description: Success
      tags:
        - Cart
      security:
        - api_key: [ 'cart-manage' ]

    delete:
      operationId: "remove-product"
      description: Remove a product from the Cart
      parameters:
        - in: path
          name: merchant-id
          description: ID of the merchant
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: product-id
          description: ID of the product
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Success
        403:
          description: The caller is not authorized to call this endpoint
        409:
          description: Error
      tags:
        - Cart
      security:
        - api_key: [ 'cart-manage' ]


components:
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: Authorization

  schemas:
    Product:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: ID of the product
        quantity:
          type: integer
          description: Number of items
          default: 1
        created:
          type: string
          format: date-time
          description: Creation Date/Time
        updated:
          type: string
          format: date-time
          description: Last modification Date/Time

    Cart:
      type: object
      properties:
        merchantId:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        created:
          type: string
          format: date-time
          description: Creation Date/Time
        updated:
          type: string
          format: date-time
          description: Last modification Date/Time

    GetCartResponse:
      type: object
      properties:
        cart:
          type: object
          $ref: "#/components/schemas/Cart"

    AddProductRequest:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: ID of the product
        quantity:
          type: integer
          description: Number of items
          default: 1
          minimum: 1

    UpdateProductRequest:
      type: object
      properties:
        quantity:
          type: integer
          description: Number of items
          default: 1
          minimum: 1

    SearchCartRequest:
      type: object
      properties:
        merchantId:
          type: integer
          format: int64
          nullable: true
        accountId:
          type: integer
          format: int64
          nullable: true
        limit:
          type: integer
          default: 30
        offset:
          type: integer
          default: 0

    SearchCartResponse:
      type: object
      properties:
        carts:
          type: array
          items:
            $ref: "#/components/schemas/CartSummary"

    CartSummary:
      type: object
      properties:
        merchantId:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        created:
          type: string
          format: date-time
          description: Creation Date/Time
        updated:
          type: string
          format: date-time
          description: Last modification Date/Time