# src/main/kotlin/com/wutsi/ecommerce/cart/dto/CartSummary.kt
package com.wutsi.ecommerce.cart.dto

import org.springframework.format.`annotation`.DateTimeFormat
import java.time.OffsetDateTime
import kotlin.Long

public data class CartSummary(
    public val merchantId: Long = 0,
    public val accountId: Long = 0,
    @get:DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ssZ")
    public val created: OffsetDateTime = OffsetDateTime.now(),
    @get:DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ssZ")
    public val updated: OffsetDateTime = OffsetDateTime.now()
)